+++ src/pokemon.c
@@ -3095,9 +3095,9 @@ s32 CalculateBaseDamage(struct BattlePokemon *attacker, struct BattlePokemon *de
     u16 attack, defense;
     u16 spAttack, spDefense;
     u8 defenderHoldEffect;
-    u8 defenderHoldEffectParam;
+    u16 defenderHoldEffectParam;
     u8 attackerHoldEffect;
-    u8 attackerHoldEffectParam;
+    u16 attackerHoldEffectParam;
 
     bool8 hasStab;
     bool8 higherAtt;
